+++ packages/react-native-executorch/common/rnexecutorch/data_processing/dsp.h
@@ -5,7 +5,7 @@
 
 namespace rnexecutorch::dsp {
 
-std::vector<float> hannWindow(int size);
+std::vector<float> hannWindow(size_t size);
 std::vector<float> stftFromWaveform(std::span<float> waveform,
                                     size_t fftWindowSize, size_t hopSize);
